+++ capuchin/Delectus/DelectusPrintView.m
@@ -27,7 +27,7 @@
         selectedRange = [self selectedRange];
         [self insertText:newline replacementRange:selectedRange];
         NSArray* cols = [dataSource collectColumns];
-        int colcount = [cols count];
+        NSUInteger colcount = [cols count];
         if(colcount>0){
             int rowcount = [dataSource countRows];
             if(rowcount>0){
